+++ assets/scripts/leftText.gd
@@ -1,7 +1,7 @@
 extends RichTextLabel
 
 func _ready():
-	text = 'Earth Defence Force Terminal v'+Global.GAME_VERSION
+	text = 'Earth Defence Force Terminal v'+ProjectSettings.get_setting("application/config/version")+'\nRunning on Godot v'+Engine.get_version_info()["string"]
 	Global.DIRECTORY = Global.DIRECTORY_HOME
 
 func _unhandled_input(event):